rename: WeaviateHandler 로 네이밍 변경 - #30 -  WeaviateService -> WeaviateHandler 로 수정
