+++ src/main/java/org/example/capstonedesign1/domain/admin/service/AdminCommandService.java
@@ -4,7 +4,7 @@ import org.example.capstonedesign1.domain.user.entity.User;
 import org.example.capstonedesign1.domain.user.entity.enums.Role;
 import org.example.capstonedesign1.global.exception.ForbiddenException;
 import org.example.capstonedesign1.global.exception.code.ErrorCode;
-import org.example.capstonedesign1.global.weaviate.service.WeaviateService;
+import org.example.capstonedesign1.global.weaviate.WeaviateHandler;
 import org.springframework.stereotype.Service;
 
 import lombok.RequiredArgsConstructor;
@@ -15,22 +15,22 @@ import lombok.extern.log4j.Log4j2;
 @RequiredArgsConstructor
 public class AdminCommandService {
 
-	private final WeaviateService weaviateService;
+	private final WeaviateHandler weaviateHandler;
 
 	public void createSchema(User user) {
 		if (!user.getRole().equals(Role.ROLE_ADMIN)) {
 			throw new ForbiddenException(ErrorCode.UN_AUTHORIZED);
 		}
-		weaviateService.createBankProductSchema();
-		weaviateService.createCardProductSchema();
+		weaviateHandler.createBankProductSchema();
+		weaviateHandler.createCardProductSchema();
 	}
 
 	public void setData(User user) {
 		if (!user.getRole().equals(Role.ROLE_ADMIN)) {
 			throw new ForbiddenException(ErrorCode.UN_AUTHORIZED);
 		}
-		weaviateService.setCardProductDataFromDB();
-		weaviateService.setBankProductDataFromDB();
+		weaviateHandler.setCardProductDataFromDB();
+		weaviateHandler.setBankProductDataFromDB();
 	}
 
 }
